Rewrote base word with database
Rewrote base word with database
